\name{dataset1}
\alias{dataset1}
\docType{data}
\title{
   Dataset for MVN test
}
\description{
  Dataset built for MVN mixture test, which contains 3 variables and 25 observations.
}
\usage{data("dataset1")}
\format{
  A data frame with 25 observations on 3 independent variables, named as \code{fac1}, \code{fac2} and \code{fac3}.
  \describe{
    \item{\code{fac1}}{The 1st factor.}
    \item{\code{fac2}}{The 2nd factor.}
    \item{\code{fac3}}{The 3rd factor.}
  }
}
\examples{
  dataset1
}
\keyword{datasets}
\keyword{MVN distribution}
