\name{dataset2}
\alias{dataset2}
\docType{data}
\title{
   dataset2
}
\description{
  dataset built for multivariate normal mixture distribution
}
\usage{data("dataset2")}
\format{
  A data frame with 96 pseudo-observations generated by random number generator. All observations come from 3 different centers which have been marked in the last column "species". More specifically, data of species=1 comes from the center (1,1,1,1); data of species=2 comes from the center (2,2,2,0);  data of species=2 comes from the center (1,0,2,2).
  \describe{
    \item{\code{dimen1}}{the 1st dimension of observation vectors}
    \item{\code{dimen2}}{the 2nd dimension of observation vectors}
    \item{\code{dimen3}}{the 3rd dimension of observation vectors}
    \item{\code{dimen4}}{the 4th dimension of observation vectors}
    \item{\code{species}}{The columns to mark down species of each observation, supporting reference for further clustering task}
  }
}
\examples{
dataset2
}
\keyword{datasets}
