\name{dataset1}
\alias{dataset1}
\docType{data}
\title{
   dataset1
}
\description{
  dataset built for multivariate normal distribution
}
\usage{data("dataset1")}
\format{
  A data frame with 25 observations on 3 different alloy contents (factors or variables), represented by fac1, fac2 and fac3.
  \describe{
    \item{\code{fac1}}{The 1st factor}
    \item{\code{fac2}}{The 2nd factor}
    \item{\code{fac3}}{The 3rd factor}
  }
}
\examples{
dataset1
}
\keyword{datasets}
