\name{dataset2}
\alias{dataset2}
\docType{data}
\title{
   Dataset for MVN mixture test
}
\description{
  Dataset built for MVN mixture test, which contains 4 variables (the first 4 columns), clustering (the last column) and 96 observations.
}
\usage{data("dataset2")}
\format{
  A data frame with 96 pseudo-observations generated by random number generator. All observations come from 3 different centers which have been marked in the last column "species". More specifically, data of species=1 comes from the center (1,1,1,1); data of species=2 comes from the center (2,2,2,0); data of species=3 comes from the center (1,0,2,2).
  \describe{
    \item{\code{dimen1}}{the 1st variable}
    \item{\code{dimen2}}{the 2nd variable}
    \item{\code{dimen3}}{the 3rd variable}
    \item{\code{dimen4}}{the 4th variable}
    \item{\code{species}}{clustering label}
  }
}
\examples{
dataset2
}
\keyword{datasets}
\keyword{MVN mixture distribution}
