\name{MVN_BayesianIterator}
\alias{MVN_BayesianIterator}
\title{
  MVN_BayesianIterator
}
\description{
  Function to execute precise parameter estimation for multivariate normal distribution under Bayesian framework.
}
\usage{
# Get parameters of Bayesian posterior MVN:
MVN_BayesianIterator(data, pri_mean=colMeans(data), Gibbs_nums=5000,
pseudo_nums=dim(data)[1], threshold=1e-04, iteration=100, ...)
}
\arguments{
  \item{data}{
     Design matrix: data.frame or matrix-like data, \code{dim(data)[1]} should be the number of observations and \code{dim(data)[2]} should be the number of variables or parameters.
}
  \item{pri_mean}{
     Priori mean: necessary vector which should be of the identical dimensions of data (\code{length(pri_mean) == dim(data)[2]}). Default value uses mean vector of data.
}
  \item{Gibbs_nums}{
     Arg to determine numbers of Gibbs sampling for each step. the defult value uses 5000.
}
  \item{pseudo_nums}{
     Arg to control how many generated samples are used in iteration step. Default value keeps as the same as that of input data. Notice that too small settings can result in error.
}
  \item{threshold}{
     Paramter of condition for controlling stoping the iteration loop. Default value used 0.0001. When the Euclidean distance between mean of Gibbs samplings (last \code{psedu_nums} items) and Bayesian posterior mean is less than that value, iteration stops.
}
  \item{iteration}{
     Maximum steps for iteration. Default value = 100. If the loop does not converge after 100-th step, it will automatically exit.
}
  \item{\dots}{
     Other arguments to control the process of Gibbs sampling.
}
}
\details{
  It can be proved that if the sampling distritution comes from a multivariate normal distribution, another multivariate normal distribution can always be the conjugate prior distribution of the sampling distribution. Because of that property, the convergence of Beyesian iterator for multivariate normal distribution can be ensured. But it needs to be pointed out that parameter of \code{pseudo_nums} leads to the randomness of the final result. Therefore a proper setting is necessary.
}
\value{
return a list of Bayesian posterior after iteration processing:
\item{mean}{mean vector of Bayesian posterior}
\item{var}{covariance of Bayesian posterior}
}
\author{
  ZHANG Chen
}
\note{
  If the real values of parameters are the only interested thing we concerned, this iterator makes sense. It can significantly help us decrease the scale of covariance matrix, to obtain a more reliable estimation for mean values. However, in more cases, we need to discuss the correlationship between a specific combination of some parameters. These relationships are usually clued by the covariance matrix.
}
\seealso{
 MVN_BayesianPosterior, MVN_GibbsSampler, MVN_FConditional, MatrixAlternative
}
\examples{
library(mvtnorm)

# Bayesian posterior before iteration using dataset1 as example,
# c(80, 16, 3) as priori mean:
# View 2-norm of covariance matrix of Bayesian posterior:
BPos_init <- MVN_BayesianPosterior(dataset1, c(80,16,3))
BPos_init
norm(as.matrix(BPos_init$var), type = "2")

# Bayesian posterior after iteration using c(80,16,3) as priori
# Using 30 last samples generated by GibbsSampler for each step:
BPos_fina1 <- MVN_BayesianIterator(dataset1, c(80,16,3), 5000, 30)
BPos_fina1
norm(as.matrix(BPos_fina1$var), type = "2")

## Too small setting of pseudo_nums can results in singular system:
## try:
#  MVN_BayesianIterator(dataset1, pseudo_nums=3)
}
\keyword{ multivariate normal distribution }
\keyword{ Bayesian posterior }
\keyword{ Gibbs sampling }
