\name{MVN_GibbsSampler}
\alias{MVN_GibbsSampler}
\title{
  MVN_GibbsSampler
}
\description{
  Gibbs sampler for generating random vectors according to a given multivariate normal distribution
}
\usage{
# Bayesian posterior as data
# data <- MVN_BayesianPosterior(dataset1)

# Using Gibbs sampler to generate random vectors based on Bayesian posterior:
MVN_GibbsSampler(n, data, initial, reject_rate, burn)
}
\arguments{
  \item{n}{
     Numbers of generated random vectors
}
  \item{data}{
     A bivariate list which contains the mean vector (\code{data$mean}) and the covariance matrix (\code{data$var}) of the given multivariate normal distribution.
}
  \item{initial}{
     Initial value where Markov chain starts. Default value use a random vector near the mean vector generated by \code{rmvnorm()}.
}
  \item{reject_rate}{
     An argument to control burn-in period (by ratio). Default value is 0.2, meaning 20\% generated vectors in beginning will be rejected. If user customized this arg, the next arg \code{burn} should use the default value.
}
  \item{burn}{
     Another argument to control burn-in period (by numbers). If user customized this arg, the finial result will drop the same number of vectors in beginning. When \code{reject_rate} and \code{burn} were configured simultaneously, result will be generated by this method.
}
}
\details{
  There're also some literatures recommand using the mean or mode of priori as initial value. users can make their own settings according to their own needs.
}
\value{
  return a certain number of random vectors based on specified MVN distribution
}
\author{
  ZHANG Chen
}
\seealso{
 MVN_FConditional, MatrixAlternative
}
\examples{
library(mvtnorm)

# Get parameters of Bayesian posterior multivariate normal distribution
BPos <- MVN_BayesianPosterior(dataset1)
BPos

# Using previous result (BPos) to generate random vectors through Gibbs
# sampling: 7000 observations, start from c(1,1,2), use 0.3 burning rate
BPos_Gibbs <- MVN_GibbsSampler(7000, BPos, initial=c(1,1,2), 0.3)
tail(BPos_Gibbs)

# Check for convergence of Markov chain
BPos$mean
colMeans(BPos_Gibbs)
BPos$var
var(BPos_Gibbs)

## Visulization:
#library(rgl)
#fac1 <- BPos_Gibbs[,1]
#fac2 <- BPos_Gibbs[,2]
#fac3 <- BPos_Gibbs[,3]
#plot3d(x=fac1, y=fac2, z=fac3, col="red", size=2)
}
\keyword{ multivariate normal distribution }
\keyword{ Gibbs sampling }
